pyproject.toml
[project]
name = "realsense-pose"
version = "0.1.0"
description = "Add a short description here"
dependencies = [
    "pyrealsense2-macosx",
    "numpy~=1.26.2",
    "opencv-python~=4.7.0.72",
    "matplotlib~=3.8.2",
    "seaborn>=0.13.0",
    "pip~=23.3.1",
    "japanize-matplotlib~=1.1.3",
    "PyYAML>=6.0",
    "dataclass-wizard>=0.22.2",
    "dataclasses_json>=0.6.3",
    "pandas>=2.1.4",
    "jinja2>=3.1.2",
]
readme = "README.md"
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ipykernel~=6.23.1",
    "mypy~=1.3.0",
    "requests~=2.31.0",
    "black>=23.11.0",
    "snakeviz>=2.2.0",
    "types-PyYAML>=6.0.12.12",
    "types-requests>=2.31.0.10",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.pyright]
reportMissingTypeStubs = "none"